serving/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch >= 1.3
evaluate==0.4.0
transformers==4.32.0
datasets==2.14.4
typer==0.9.0
accelerate==0.22.0
rich==13.5.2
python-dotenv==1.0.0
fastapi==0.101.1
clearml==1.12.2
uvicorn==0.23.2

gradio==3.41.1
locust==2.16.1
